temporary data seeds

+++ db/migrations/20180920195533_menu_items.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('menu_items', function (t) {
     t.increments('id').primary();
     t.string('name');
-    t.integer('price');
+    t.decimal('price');
   })
 };
